Add base and board visually, add Game object
Adds a base and wireframe board to assist in gameplay. Adds "Game" object to be used to clear rows, award points, and keep track of tetromino and cube positions. Added initial X and Z position values in globals in order to line pieces up with the board and keep the board centered on x = 0, z = 0.